refactor: remove unused components, duplicate utilities, and unused dependencies
