+++ backend/src/services/runtime.ts
@@ -5,11 +5,6 @@ import {
   createMetricsRegistry,
   type MetricsRegistry,
 } from "../lib/observability/metrics";
-import {
-  configureAlertingService,
-  createAlertingService,
-  type AlertingService,
-} from "../lib/observability/alerting";
 import { configureSubscriptionService } from "../modules/billing/subscription-service";
 import {
   getStripeClient,
@@ -24,7 +19,6 @@ export interface RuntimeServices {
   db: Database;
   cacheService: CacheService;
   metrics: MetricsRegistry;
-  alerting: AlertingService;
   stripe: ReturnType<typeof getStripeClient> | null;
   email: ReturnType<typeof createEmailService> | null;
 }
@@ -33,7 +27,6 @@ export function createRuntimeServices(db: Database): RuntimeServices {
   const config = getConfig();
   const cacheService = createCacheService();
   const metrics = createMetricsRegistry();
-  const alerting = createAlertingService();
   const stripe = config.BILLING_MODE === "managed" ? getStripeClient() : null;
   const email =
     config.EMAIL_MODE !== "disabled"
@@ -41,7 +34,6 @@ export function createRuntimeServices(db: Database): RuntimeServices {
       : null;
 
   configureMetricsRegistry(metrics);
-  configureAlertingService(alerting);
   configureSubscriptionService({ db, cacheService });
   if (email) {
     configureEmailService(email);
@@ -51,7 +43,6 @@ export function createRuntimeServices(db: Database): RuntimeServices {
     db,
     cacheService,
     metrics,
-    alerting,
     stripe,
     email,
   };
