--- frontend/run-test-debug.ts
@@ -1,1 +1,0 @@
-// can't run this directly because it's not exported. Let me just add a console.log inside the test to see what totalsDuringMutation is.
