--- frontend/src/components/ui/SavedMealButton.tsx
@@ -1,62 +1,0 @@
-import { useState } from "react";
-
-import { useCreateSavedMeal } from "@/hooks/queries/useSavedMeals";
-import { useStore } from "@/store/store";
-
-import IconButton from "./IconButton";
-import { StarIcon } from "./Icons";
-
-interface SavedMealButtonProps {
-  mealName: string;
-  protein: number;
-  carbs: number;
-  fats: number;
-  mealType?: "breakfast" | "lunch" | "dinner" | "snack";
-  ingredients?: unknown[];
-  ariaLabel?: string;
-}
-
-export default function SavedMealButton({
-  mealName,
-  protein,
-  carbs,
-  fats,
-  mealType = "snack",
-  ingredients,
-  ariaLabel = "Save meal for quick re-entry",
-}: SavedMealButtonProps) {
-  const [isSaved, setIsSaved] = useState(false);
-  const createSavedMeal = useCreateSavedMeal();
-  const { showNotification } = useStore();
-
-  const handleSave = async () => {
-    try {
-      await createSavedMeal.mutateAsync({
-        name: mealName || "Saved Meal",
-        protein,
-        carbs,
-        fats,
-        mealType,
-        ingredients,
-      });
-      setIsSaved(true);
-      showNotification("Meal saved for quick re-entry", "success");
-    } catch (error) {
-      showNotification(
-        error instanceof Error ? error.message : "Failed to save meal",
-        "error",
-      );
-    }
-  };
-
-  return (
-    <IconButton
-      variant="custom"
-      className={isSaved ? "text-primary" : "text-muted hover:text-foreground"}
-      onClick={handleSave}
-      disabled={isSaved || createSavedMeal.isPending}
-      ariaLabel={ariaLabel}
-      icon={<StarIcon className={`h-4 w-4 ${isSaved ? "fill-current" : ""}`} />}
-    />
-  );
-}
