--- frontend/src/components/ui/VirtualList.tsx
@@ -1,65 +1,0 @@
-import { type ReactNode,useRef } from "react";
-import { useVirtualizer } from "@tanstack/react-virtual";
-
-interface VirtualListProps<T> {
-  items: T[];
-  renderItem: (item: T, index: number) => ReactNode;
-  estimateSize?: number;
-  overscan?: number;
-  className?: string;
-  emptyState?: ReactNode;
-}
-
-export function VirtualList<T>({
-  items,
-  renderItem,
-  estimateSize = 80,
-  overscan = 5,
-  className = "",
-  emptyState,
-}: VirtualListProps<T>) {
-  const parentReference = useRef<HTMLDivElement>(null);
-
-  const virtualizer = useVirtualizer({
-    count: items.length,
-    getScrollElement: () => parentReference.current,
-    estimateSize: () => estimateSize,
-    overscan,
-  });
-
-  if (items.length === 0 && emptyState) {
-    return <div className={className}>{emptyState}</div>;
-  }
-
-  return (
-    <div
-      ref={parentReference}
-      className={`overflow-auto ${className}`}
-      style={{ height: "100%", width: "100%" }}
-    >
-      <div
-        style={{
-          height: `${virtualizer.getTotalSize()}px`,
-          width: "100%",
-          position: "relative",
-        }}
-      >
-        {virtualizer.getVirtualItems().map((virtualItem) => (
-          <div
-            key={virtualItem.key}
-            style={{
-              position: "absolute",
-              top: 0,
-              left: 0,
-              width: "100%",
-              height: `${virtualItem.size}px`,
-              transform: `translateY(${virtualItem.start}px)`,
-            }}
-          >
-            {renderItem(items[virtualItem.index], virtualItem.index)}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
