--- frontend/src/features/goals/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useEditableMacroTarget } from "./useMacroTarget";
