+++ frontend/src/features/goals/hooks/page/index.ts
@@ -1,4 +1,3 @@
-export * from "./useGoalsActions";
 export * from "./useGoalsData";
 export * from "./useGoalsMutations";
 export * from "./useGoalsPage";
