+++ frontend/src/features/reporting/utils/index.ts
@@ -1,7 +1,7 @@
 // Export all utility functions for easy importing
 export * from "./dataProcessing";
-export * from "./dateUtilities";
 export * from "./macroCalculations";
+export * from "@/utils/dateUtilities";
 
 // Re-export the main insights calculations for backward compatibility
 export * from "./insightsCalculations";
