--- frontend/src/features/settings/utils/constants.ts
@@ -1,3 +1,0 @@
-// Import directly from @/utils/userConstants
-// This file is kept for backwards compatibility but will be removed
-export * from "@/utils/userConstants";
