+++ frontend/src/services/native/index.ts
@@ -1,5 +1,3 @@
 export * from "./appLifecycle";
-export * from "./camera";
-export * from "./haptics";
 export * from "./platform";
 export * from "./statusBar";
