+++ frontend/src/utils/dateUtilities.ts
@@ -160,3 +160,22 @@ export function getDateRangeData(
     days,
   };
 }
+
+export const formatDate = formatDateShort;
+
+export function getDayString(date: Date) {
+  return `${date.getFullYear()}-${(date.getMonth() + 1).toString().padStart(2, "0")}-${date.getDate().toString().padStart(2, "0")}`;
+}
+
+export function getWeekString(date: Date) {
+  const year = date.getFullYear();
+  const firstDay = new Date(date.getFullYear(), 0, 1);
+  const pastDaysOfYear = (date.getTime() - firstDay.getTime()) / 86_400_000;
+  const week = Math.ceil((pastDaysOfYear + firstDay.getDay() + 1) / 7);
+
+  return `${year}-W${week.toString().padStart(2, "0")}`;
+}
+
+export function getMonthString(date: Date) {
+  return `${date.getFullYear()}-${(date.getMonth() + 1).toString().padStart(2, "0")}`;
+}
